requirements.txt
playwright==1.41.0
gspread==6.0.0
google-auth==2.27.0
python-dotenv==1.0.0
PyYAML==6.0.1
